requirements.txt
accelerate==1.12.0
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
airportsdata==20250811
annotated-types==0.7.0
anthropic==0.64.0
anyio==4.10.0
argcomplete==3.6.2
astor==0.8.1
asttokens==3.0.0
attrs==25.3.0
av @ https://image-url-2-feature-1251524319.cos.ap-shanghai.myqcloud.com/tuxi1/av-12.0.0rc1-cp312-cp312-manylinux_2_31_x86_64.whl
black==25.1.0
blake3==1.0.5
blinker==1.4
blobfile==3.0.0
boto3==1.33.13
botocore==1.33.13
build==1.3.0
cachetools==6.2.0
cbor2==5.7.0
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
coloredlogs==15.0.1
compressed-tensors==0.11.0
confluent-kafka==2.11.1
contourpy==1.3.3
cos_python_sdk_v5==1.9.38
coscmd @ https://image-url-2-feature-1251524319.cos.ap-shanghai.myqcloud.com/tuxi1/coscmd-1.9.0.2-py3-none-any.whl
crcmod==1.7
cryptography==3.4.8
cuda-bindings==12.9.2
cuda-pathfinder==1.1.0
cuda-python==12.9.0
cupy-cuda12x==13.6.0
cvcuda-cu12==0.15.0
cycler==0.12.1
Cython==3.1.4
dashscope==1.25.8
datamodel-code-generator==0.33.0
datasets==4.0.0
DateTime==5.5
dbus-python==1.2.18
decorator==5.2.1
decord==0.6.0
deep_ep @ file:///sgl-workspace/DeepEP
depyf==0.19.0
devscripts===2.22.1ubuntu1
dill==0.3.8
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
distro-info==1.1+ubuntu0.2
dnspython==2.8.0
einops==0.8.1
email-validator==2.3.0
executing==2.2.0
fakeredis==2.31.1
fastapi==0.116.1
fastapi-cli==0.0.12
fastapi-cloud-cli==0.1.5
fastrlock==0.8.3
filelock==3.19.1
flash_attn @ https://image-url-2-feature-1251524319.cos.ap-shanghai.myqcloud.com/tuxi1/flash_attn-2.8.3%2Bcu12torch2.8cxx11abiTRUE-cp312-cp312-linux_x86_64.whl
flashinfer-python==0.3.1
flatbuffers==25.2.10
fonttools==4.59.2
fpng_py==0.0.3
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.3.0
future==1.0.0
genson==1.3.0
gguf==0.17.1
gunicorn==22.0.0
h11==0.16.0
hf-xet==1.1.8
hf_transfer==0.1.9
html5lib==1.1
httpcore==1.0.9
httplib2==0.20.2
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
icdiff==2.0.7
identify==2.6.13
idna==3.10
importlib-metadata==4.6.4
inflect==7.5.0
iniconfig==2.1.0
interegular==0.3.3
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kafka-python-ng==2.2.3
keyring==23.5.0
kiwisolver==1.4.9
lark==1.2.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.11.3
lxml==6.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.8.5
ml_dtypes==0.5.3
modelscope==1.29.1
mooncake-transfer-engine==0.3.5
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
nixl==0.5.0
nodeenv==1.9.1
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-frontend==1.14.0
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.27.3
nvidia-nvimgcodec-cu12==0.6.0.32
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.0
onnx==1.19.0
onnxruntime-gpu==1.22.0
openai==1.99.1
openai-harmony==0.0.4
opencv-contrib-python-headless==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.11.3
outlines==0.1.11
outlines_core==0.2.11
packaging==25.0
pandas==2.3.2
parso==0.8.5
partial-json-parser==0.2.1.1.post6
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
pillow_heif==0.21.0
platformdirs==4.4.0
pluggy==1.6.0
ply==3.11
pre_commit==4.3.0
prettytable==3.16.0
prometheus-async==22.2.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.20.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.1
pyarrow==21.0.0
pybase64==1.4.2
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.23.0
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic_core==2.33.2
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.10.1
pynvml==12.0.0
pyparsing==2.4.7
pyPEG2==2.15.2
pyproject_hooks==1.2.0
pytest==8.4.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
PyTurboJPEG==1.7.7
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.2
qwen-vl-utils==0.0.14
ray==2.49.1
redenv==0.0.5
redeuler==1.4.3
redis==5.3.1
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rocketmq==0.4.4
rpds-py==0.27.1
s3transfer==0.8.2
safetensors==0.6.2
scikit_build_core==0.11.6
scipy==1.16.1
SecretStorage==3.3.1
sentencepiece==0.2.1
sentry-sdk==2.38.0
setproctitle==1.3.6
setuptools==79.0.1
sgl-kernel==0.3.7
-e git+https://github.com/sgl-project/sglang.git@bc80dc4ce0ae5a97b8a58faa1d8b8cfbb56e21f5#egg=sglang&subdirectory=python
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==1.0.0
stack-data==0.6.3
starlette==0.47.3
sympy==1.14.0
tabulate==0.9.0
tensorrt @ file:///opt/tensorrt/python/tensorrt-10.12.0.36-cp312-none-linux_x86_64.whl#sha256=7ef60e4f20d958cc678ef444f79c593605a8f03ff6be669f8102606cccb35309
tensorrt_dispatch @ file:///opt/tensorrt/python/tensorrt_dispatch-10.12.0.36-cp312-none-linux_x86_64.whl#sha256=feafc09e869a645a1851dde1ab7e657d8222524d4a946e15bc89af33630903f6
tensorrt_lean @ file:///opt/tensorrt/python/tensorrt_lean-10.12.0.36-cp312-none-linux_x86_64.whl#sha256=d0038a24ad0302756a8e9be1e548c233a66ef1bdf78bc339651a24ff3714f55d
thrift==0.11.0
thriftpy2==0.5.2
tiktoken==0.11.0
timm==1.0.16
tokenizers==0.22.2
torch==2.8.0+cu126
torch_memory_saver==0.0.8
torchao==0.9.0+cu126
torchaudio==2.8.0+cu126
torchvision==0.23.0+cu126
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
triton==3.4.0
typeguard==4.4.4
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
unattended-upgrades==0.1
urllib3==2.0.7
uv==0.8.18
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.34.0
vllm @ https://image-url-2-feature-1251524319.cos.ap-shanghai.myqcloud.com/tuxi1/vllm-0.11.0-cp38-abi3-manylinux1_x86_64.whl
wadllib==1.3.6
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
webp @ https://image-url-2-feature-1251524319.cos.ap-shanghai.myqcloud.com/tuxi1/webp-0.4.4-cp38-abi3-linux_x86_64.whl
websocket-client==1.9.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
xformers==0.0.32.post1
xgrammar==0.1.25
xmltodict==1.0.2
xxhash==3.5.0
yapf==0.43.0
yarl==1.20.1
zipp==1.0.0
zope.interface==8.0
